Refactor component props declaration
