--- src/components/AnswerReaction/type.ts
@@ -1,11 +1,0 @@
-import Position from "../../types/position.type";
-import { VersionData } from "../../types/pokemonData.type";
-
-type AnswerReactionProps = {
-  isCorrect: boolean;
-  imagePosition: Position;
-  clickPosition: Position;
-  gameData: VersionData;
-};
-
-export default AnswerReactionProps;
