--- src/components/GameVersionCard/type.ts
@@ -1,11 +1,0 @@
-import Difficulty from "../../types/difficulty.type";
-import GameVersion from "../../types/gameVersion.type";
-import { VersionData } from "../../types/pokemonData.type";
-
-type GameVersionCardProps = {
-  difficulty: Difficulty;
-  gameData: VersionData;
-  gameVersion: GameVersion;
-};
-
-export default GameVersionCardProps;
