--- src/components/NavBar/type.ts
@@ -1,10 +1,0 @@
-import GameProgress from "../../types/gameProgress.type";
-import { VersionData } from "../../types/pokemonData.type";
-
-type NavBarProps = {
-  gameData: VersionData;
-  gameProgress: GameProgress;
-  elapsedTime: string;
-};
-
-export default NavBarProps;
