--- src/components/PokemonNameList/type.ts
@@ -1,9 +1,0 @@
-import { MouseEvent } from "react";
-import { VersionData } from "../../types/pokemonData.type";
-
-type PokemonNameListProps = {
-  gameData: VersionData;
-  handlePickedOption?: (e: MouseEvent, pickedPokemon: string) => void;
-};
-
-export default PokemonNameListProps;
