+++ src/tests/components/AnswerReaction.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import AnswerReaction from "../../components/AnswerReaction/AnswerReaction";
+import AnswerReaction from "../../components/AnswerReaction";
 import "@testing-library/jest-dom";
 
 const defaultVisualViewport = {
