+++ src/tests/components/NavBar.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
-import NavBar from "../../components/NavBar/NavBar";
+import NavBar from "../../components/NavBar";
 import "@testing-library/jest-dom";
 
 const gameData = {
